refactor(database): move MockGeoLocationService into its proper file
